+++ RELEASENOTES.md
@@ -6,6 +6,7 @@
 * Dialbacks are now included in the call detail reports [#1409]
 * HTML formatted notifications for voicemail.
 * Added enabled field to volunteer CSV and JSON exports, with download buttons on Volunteers page [#1411]
+* Root service body admins can now access meeting request logs that have no service_body_id assigned [#1399]
 
 ### 4.5.0 (July 11, 2025)
 * Added new feature that allow for creating a custom prompt for language selection feature. [#1228]